requirements.txt
Flask==2.0.1
gunicorn==20.1.0
werkzeug==2.0.3
langchain
langchain-openai
langchain_core
langchain-community
langchain-elasticsearch
langgraph
llama-index
llama-index-vector-stores-elasticsearch
openai
python-dotenv
pydantic
psycopg2-binary

